python/gerenciamento_usuarios/pyproject.toml
[tool.poetry]
name = "gerenciamento-usuarios"
version = "0.1.0"
description = ""
authors = ["FrancisLauriano <130801505+FrancisLauriano@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "9.0.0"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
